pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=64", "setuptools_scm>=8"]

[project]
authors = [
]
# Full list: https://pypi.python.org/pypi?%3Aaction=list_classifiers
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python",
]
dependencies = [
    "fastapi==0.115.4",
    "llama-index==0.11.20",
    "llama-index-llms-litellm==0.2.2",
    "nest-asyncio==1.6.0",
    "paper-qa>=5",
    "python-dotenv==1.0.1",
    "supabase==2.9.1",
    "uvicorn==0.32.0",
]
description = "Server for health-insurance-sg"
dynamic = ["version"]
keywords = []
license = {file = "LICENSE"}
maintainers = [
]
name = "health-insurance-sg-server"
readme = "README.md"
requires-python = ">=3.11"

[project.scripts]
pqa = "paperqa.agents:main"
